switch to SFMTRand

+++ src/game/Anticheat/WardenAnticheat/WardenScan.cpp
@@ -164,7 +164,7 @@ WindowsModuleScan::WindowsModuleScan(std::string const& module, bool wanted, std
     [this](Warden const* warden, std::vector<std::string>&, ByteBuffer& scan)
     {
         auto const winWarden = reinterpret_cast<WardenWin const*>(warden);
-        auto const seed = static_cast<uint32>(rand32());
+        auto const seed = rand32();
 
         scan << static_cast<uint8>(winWarden->GetModule()->opcodes[FIND_MODULE_BY_NAME] ^ winWarden->GetXor()) << seed;
 
@@ -192,7 +192,7 @@ WindowsModuleScan::WindowsModuleScan(std::string const& module, CheckT checker, 
     [this](Warden const* warden, std::vector<std::string>&, ByteBuffer& scan)
     {
         auto const winWarden = reinterpret_cast<WardenWin const*>(warden);
-        auto const seed = static_cast<uint32>(rand32());
+        auto const seed = rand32();
 
         scan << static_cast<uint8>(winWarden->GetModule()->opcodes[FIND_MODULE_BY_NAME] ^ winWarden->GetXor()) << seed;
 
@@ -320,7 +320,7 @@ WindowsCodeScan::WindowsCodeScan(uint32 offset, std::vector<uint8> const& patter
     [this](Warden const* warden, std::vector<std::string>&, ByteBuffer& scan)
     {
         auto const winWarden = reinterpret_cast<WardenWin const*>(warden);
-        auto const seed = static_cast<uint32>(rand32());
+        auto const seed = rand32();
 
         scan << static_cast<uint8>(winWarden->GetModule()->opcodes[this->m_memImageOnly ? FIND_MEM_IMAGE_CODE_BY_HASH : FIND_CODE_BY_HASH] ^ winWarden->GetXor())
              << seed;
@@ -475,7 +475,7 @@ WindowsHookScan::WindowsHookScan(std::string const& module, std::string const& p
         strings.emplace_back(this->m_proc);
 
         auto const winWarden = reinterpret_cast<WardenWin const*>(warden);
-        auto const seed = static_cast<uint32>(rand32());
+        auto const seed = rand32();
 
         scan << static_cast<uint8>(winWarden->GetModule()->opcodes[API_CHECK] ^ winWarden->GetXor()) << seed;
 
@@ -507,7 +507,7 @@ WindowsDriverScan::WindowsDriverScan(std::string const& name, std::string const&
         strings.emplace_back(this->m_name);
 
         auto const winWarden = reinterpret_cast<WardenWin const*>(warden);
-        auto const seed = static_cast<uint32>(rand32());
+        auto const seed = rand32();
 
         scan << static_cast<uint8>(winWarden->GetModule()->opcodes[FIND_DRIVER_BY_NAME] ^ winWarden->GetXor()) << seed;
