+++ src/realmd/AuthSocket.cpp
@@ -510,7 +510,7 @@ void AuthSocket::_HandleLogonChallenge()
                     {
                         sLog.Out(LOG_BASIC, LOG_LVL_BASIC, "[AuthChallenge] Account '%s' using IP '%s' requires PIN authentication", self->m_login.c_str(), self->GetRemoteIpString().c_str());
 
-                        uint32 gridSeedPkt = self->m_gridSeed = static_cast<uint32>(rand32());
+                        uint32 gridSeedPkt = self->m_gridSeed = rand32();
                         EndianConvert(gridSeedPkt);
                         self->m_serverSecuritySalt.SetRand(16 * 8); // 16 bytes random
